Adding visual tests for lottie
